// whtree/jssdk/postgrease/src/types/oids.ts
export const DataTypeOids = {
  bool: 16,
  bytea: 17,
  char: 18,
  name: 19,
  int8: 20,
  int2: 21,
  int2vector: 22,
  int4: 23,
  regproc: 24,
  text: 25,
  oid: 26,
  tid: 27,
  xid: 28,
  cid: 29,
  oidvector: 30,
  pg_ddl_command: 32,
  pg_type: 71,
  pg_attribute: 75,
  pg_proc: 81,
  pg_class: 83,
  json: 114,
  xml: 142,
  _xml: 143,
  pg_node_tree: 194,
  _json: 199,
  _pg_type: 210,
  table_am_handler: 269,
  _pg_attribute: 270,
  _xid8: 271,
  _pg_proc: 272,
  _pg_class: 273,
  index_am_handler: 325,
  point: 600,
  lseg: 601,
  path: 602,
  box: 603,
  polygon: 604,
  line: 628,
  _line: 629,
  cidr: 650,
  _cidr: 651,
  float4: 700,
  float8: 701,
  unknown: 705,
  circle: 718,
  _circle: 719,
  macaddr8: 774,
  _macaddr8: 775,
  money: 790,
  _money: 791,
  macaddr: 829,
  inet: 869,
  _bool: 1000,
  _bytea: 1001,
  _char: 1002,
  _name: 1003,
  _int2: 1005,
  _int2vector: 1006,
  _int4: 1007,
  _regproc: 1008,
  _text: 1009,
  _tid: 1010,
  _xid: 1011,
  _cid: 1012,
  _oidvector: 1013,
  _bpchar: 1014,
  _varchar: 1015,
  _int8: 1016,
  _point: 1017,
  _lseg: 1018,
  _path: 1019,
  _box: 1020,
  _float4: 1021,
  _float8: 1022,
  _polygon: 1027,
  _oid: 1028,
  aclitem: 1033,
  _aclitem: 1034,
  _macaddr: 1040,
  _inet: 1041,
  bpchar: 1042,
  varchar: 1043,
  date: 1082,
  time: 1083,
  timestamp: 1114,
  _timestamp: 1115,
  _date: 1182,
  _time: 1183,
  timestamptz: 1184,
  _timestamptz: 1185,
  interval: 1186,
  _interval: 1187,
  _numeric: 1231,
  _cstring: 1263,
  timetz: 1266,
  _timetz: 1270,
  bit: 1560,
  _bit: 1561,
  varbit: 1562,
  _varbit: 1563,
  numeric: 1700,
  refcursor: 1790,
  _refcursor: 2201,
  regprocedure: 2202,
  regoper: 2203,
  regoperator: 2204,
  regclass: 2205,
  regtype: 2206,
  _regprocedure: 2207,
  _regoper: 2208,
  _regoperator: 2209,
  _regclass: 2210,
  _regtype: 2211,
  record: 2249,
  cstring: 2275,
  any: 2276,
  anyarray: 2277,
  void: 2278,
  trigger: 2279,
  language_handler: 2280,
  internal: 2281,
  anyelement: 2283,
  _record: 2287,
  anynonarray: 2776,
  _txid_snapshot: 2949,
  uuid: 2950,
  _uuid: 2951,
  txid_snapshot: 2970,
  fdw_handler: 3115,
  pg_lsn: 3220,
  _pg_lsn: 3221,
  tsm_handler: 3310,
  pg_ndistinct: 3361,
  pg_dependencies: 3402,
  anyenum: 3500,
  tsvector: 3614,
  tsquery: 3615,
  gtsvector: 3642,
  _tsvector: 3643,
  _gtsvector: 3644,
  _tsquery: 3645,
  regconfig: 3734,
  _regconfig: 3735,
  regdictionary: 3769,
  _regdictionary: 3770,
  jsonb: 3802,
  _jsonb: 3807,
  anyrange: 3831,
  event_trigger: 3838,
  int4range: 3904,
  _int4range: 3905,
  numrange: 3906,
  _numrange: 3907,
  tsrange: 3908,
  _tsrange: 3909,
  tstzrange: 3910,
  _tstzrange: 3911,
  daterange: 3912,
  _daterange: 3913,
  int8range: 3926,
  _int8range: 3927,
  jsonpath: 4072,
  _jsonpath: 4073,
  regnamespace: 4089,
  _regnamespace: 4090,
  regrole: 4096,
  _regrole: 4097,
  regcollation: 4191,
  _regcollation: 4192,
  int4multirange: 4451,
  nummultirange: 4532,
  tsmultirange: 4533,
  tstzmultirange: 4534,
  datemultirange: 4535,
  int8multirange: 4536,
  anymultirange: 4537,
  anycompatiblemultirange: 4538,
  pg_brin_bloom_summary: 4600,
  pg_brin_minmax_multi_summary: 4601,
  pg_mcv_list: 5017,
  pg_snapshot: 5038,
  _pg_snapshot: 5039,
  xid8: 5069,
  anycompatible: 5077,
  anycompatiblearray: 5078,
  anycompatiblenonarray: 5079,
  anycompatiblerange: 5080,
  _int4multirange: 6150,
  _nummultirange: 6151,
  _tsmultirange: 6152,
  _tstzmultirange: 6153,
  _datemultirange: 6155,
  _int8multirange: 6157,
} as const;